fix: main과 파일 병합
