+++ src/main/java/org/example/mjuteam4/comment/CommentService.java
@@ -5,8 +5,8 @@ import org.example.mjuteam4.comment.dto.request.CommentRequest;
 import org.example.mjuteam4.comment.entity.Comment;
 import org.example.mjuteam4.comment.exception.CommentNotFound;
 import org.example.mjuteam4.comment.exception.CommentAccessUnauthorized;
+import org.example.mjuteam4.member.Member;
 import org.example.mjuteam4.member.MemberService;
-import org.example.mjuteam4.member.entity.Member;
 import org.example.mjuteam4.question.QuestionService;
 import org.example.mjuteam4.question.entity.Question;
 import org.springframework.data.domain.Page;
