+++ src/main/java/org/example/mjuteam4/commentLike/CommentLikeRepository.java
@@ -3,7 +3,7 @@ package org.example.mjuteam4.commentLike;
 
 import org.example.mjuteam4.comment.entity.Comment;
 import org.example.mjuteam4.commentLike.entity.CommentLike;
-import org.example.mjuteam4.member.entity.Member;
+import org.example.mjuteam4.member.Member;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
