+++ src/main/java/org/example/mjuteam4/member/MemberRepository.java
@@ -1,8 +1,7 @@
 package org.example.mjuteam4.member;
 
-import org.example.mjuteam4.member.entity.Member;
 import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.data.repository.CrudRepository;
+
 import java.util.Optional;
 
 public interface MemberRepository extends JpaRepository<Member, Long> {
