+++ src/main/java/org/example/mjuteam4/member/security/OAuth2MemberService.java
@@ -2,11 +2,9 @@ package org.example.mjuteam4.member.security;
 
 import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
-import org.example.mjuteam4.global.exception.GlobalException;
 import org.example.mjuteam4.member.Member;
 import org.example.mjuteam4.member.MemberRepository;
 import org.example.mjuteam4.member.exception.MemberAlreadyExistsException;
-import org.example.mjuteam4.member.exception.MemberNotFoundException;
 import org.springframework.security.oauth2.client.userinfo.DefaultOAuth2UserService;
 import org.springframework.security.oauth2.client.userinfo.OAuth2UserRequest;
 import org.springframework.security.oauth2.core.OAuth2AuthenticationException;
