+++ src/main/java/org/example/mjuteam4/question/entity/Question.java
@@ -4,8 +4,8 @@ import jakarta.persistence.*;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.example.mjuteam4.member.entity.Member;
 import org.example.mjuteam4.comment.entity.Comment;
+import org.example.mjuteam4.member.Member;
 import org.example.mjuteam4.questionImage.entity.QuestionImage;
 import org.example.mjuteam4.question.dto.request.QuestionRequest;
 import org.springframework.util.StringUtils;
