+++ src/main/java/org/example/mjuteam4/tradePost/entity/TradePost.java
@@ -3,9 +3,7 @@ package org.example.mjuteam4.tradePost.entity;
 import jakarta.persistence.*;
 import lombok.Data;
 import lombok.NoArgsConstructor;
-import org.example.mjuteam4.member.entity.Member;
-import org.example.mjuteam4.question.dto.request.QuestionRequest;
-import org.example.mjuteam4.question.entity.Question;
+import org.example.mjuteam4.member.Member;
 import org.example.mjuteam4.tradePost.dto.request.TradePostRequest;
 import org.example.mjuteam4.tradePostImage.entity.TradePostImage;
 import org.springframework.util.StringUtils;
